Relax dependencies slightly
The testing harness is quite complex and does not work out of the box
as is when running on anything less than rails 5.1.

The library code has been run on rails 4, but for testing purposes when
packaged as a gem I'm not gonna lower below 5.1 unless someone posts
a pull-request with a patch to support rails 4 or 5.